+++ modules/apps/questions/questions-web/src/main/resources/META-INF/resources/js/App.es.js
@@ -33,17 +33,9 @@ export default function App(props) {
 
 	const packageName = props.npmResolvedPackageName;
 
-	const questionComponent = Liferay.FeatureFlags['LPS-167151']
-		? `${packageName}/js/pages/questions/Question.new.es`
-		: `${packageName}/js/pages/questions/Question.es`;
-
 	const questionsComponent = Liferay.FeatureFlags['LPS-165491']
 		? `${packageName}/js/pages/questions/Questions.new.es`
 		: `${packageName}/js/pages/questions/Questions.es`;
-
-	const userActivityPage = Liferay.FeatureFlags['LPS-167151']
-		? `${packageName}/js/pages/home/UserActivity.new.es`
-		: `${packageName}/js/pages/home/UserActivity.es`;
 
 	let path = props.historyRouterBasePath;
 
@@ -103,7 +95,7 @@ export default function App(props) {
 								<Route
 									component={(props) => (
 										<Component
-											module={userActivityPage}
+											module={`${packageName}/js/pages/home/UserActivity.es`}
 											props={props}
 										/>
 									)}
@@ -200,9 +192,7 @@ export default function App(props) {
 												<Route
 													component={(props) => (
 														<Component
-															module={
-																questionComponent
-															}
+															module={`${packageName}/js/pages/questions/Question.es`}
 															props={props}
 														/>
 													)}
